* I added a new implementation of Airy Ai. This implementation uses Smith's algorithm. Currently, it is provided as a separate implementation mpfr_ai2.
* Please note that both mpfr_ai and mpfr_ai2 need to efficiently evaluate
Gamma(1/3) and Gamma(2/3). This is provided by functions in the file 
gammaonethird.c

* There is no test file for mpfr_ai2 since it should quickly disappear behind
a single implementation using the best of mpfr_ai and mpfr_ai2.
However, if necessary, a test file can be obtained by copying tai.c and 
replacing mpfr_ai by mpfr_ai2 everywhere.

* I do not know if gammaonethird.c will remain like that in the future. Should
we make this implementation available as a public MPFR function? In this case,
I have to write a wrapper for providing a correctly rounded implementation.
Moreover, it contains functions that could be interesting for the 
implementation of other functions (the functions mpfr_div_ui*). Maybe, it 
could be worth providing a file specially for this purpose.

[From SVN r6804 (trunk)]