[INSTALL] Update concerning MinGW and the __USE_MINGW_ANSI_STDIO macro.
The change done in r13668 seems to be a bit wrong (at least unclear).
This was done after finding
https://fr.osdn.net/projects/mingw/lists/archive/users/2019-January/000199.html
saying that __USE_MINGW_ANSI_STDIO became deprecated, during a
discussion on the vsnprintf() behavior:
https://gmplib.org/list-archives/gmp-bugs/2020-January/004716.html

However, it seems that printf() conformity issues have completely been
resolved only recently, in version 8.
http://mingw-w64.org/doku.php says:
  Version 8 has been released
  v8.0.0: 2020-09-18
  Notable changes:
  [...]
    * __USE_MINGW_ANSI_STDIO now automatically enabled in C99 and C11
      mode when not using UCRT by Pali Rohár
  [...]

Compiling
  #include <stdio.h>
  int main (void) { volatile double d = -1.25; printf ("%e\n", d); }
gives with mingw-w64 version 7: -1.250000e+000
  and with mingw-w64 version 8: -1.250000e+00

So, when testing MPFR with mingw-w64 version 8 but with GMP still built
with mingw-w64 version 7 (and static libraries), one gets a failure in
tsprintf:
  Error in mpfr_vsprintf (s, "%e", ...);
  expected: "-1.250000e+00"
  got:      "-1.250000e+000"
because the expected result has changed (it comes from sprintf used in
the MPFR test, which is now ISO-compliant).

[From SVN r14137 (trunk)]