[tests/tget_flt.c] Removed bug_icx (duplicate test).
What was tested was just mpfr_get_flt on a NaN mpfr_t. But this was
already tested in main(). And one gets the same result and error with
icx (without -fp-model=strict). Moreover, the old test was correctly
protected by "#if !defined(MPFR_ERRDIVZERO)", needed because NaN is
defined by 0.0 / 0.0.

[From SVN r14556 (trunk)]