Add git info with "git describe" for "make check" (to be improved).
tests/Makefile.am: in output_info, add git info, but this does not
output the branch. For instance, on master, one gets something like
"4.1-root-491-g2933fad3a", and something similar on the 4.1 branch.
As the solution needs to work in detached HEAD state, we may want to
add *-branch tags to identify the first commit of a branch, and use
  git describe --match '*-branch'
instead of
  git describe --match '*-root'