[acinclude.m4] Fix/improve MPFR_CHECK_LIBM and MPFR_CHECK_LIBQUADMATH.
With a GCC snapshot, -Werror=infinite-recursion was yielding an
error in AC_CHECK_LIB for function "main". So, define MPFR_HAVE_LIB
to be used instead of AC_CHECK_LIB when it is not needed to test any
function. So, use it in MPFR_CHECK_LIBM and MPFR_CHECK_LIBQUADMATH.
Also updated MPFR_CHECK_LIBM from the current libtool.m4 code, as
done in the past.