Updated TODO to improve mpfr_pow_si/sj for negative n.
I thought that there would be some drawbacks with the underflow
by computing the power before the division, but this actually
seems to be more or less equivalent (mpfr_pow_general is already
used for MPFR_RNDN).

Also, removed the suggestion with rescaling as this may not be
possible for large values of n compared to the maximum exponent.