[tests/tsubnormal.c] Removed invalid testcase bug20220518.
The testcase bug20220518 added in
  commit 7640a832fbf0b7f9057889e076cf7f42fe4f8e58
was invalid as the mpfr_subnormalize input was not in the current
exponent range (it had exponent -1074 while emin = -1073).

As usual, inputs outside the current exponent range are not supported
(the mpfr_subnormalize description in the manual is explicit:
"The subnormal exponent range is from ‘emin’ to ‘emin+PREC(X)-1’.").
Supporting values outside the current exponent range is not needed
in practice (see examples in the manual).