[tests/cmp_str.c] In mpfr_cmp_str, check that s is a valid number.
In the tests, the string provided to mpfr_cmp_str is normally a
valid number (the common use is to check that a result is equal to
some hardcoded value given by a string). A string s that does not
represent a valid number probably has a typo, and the consequence
was that in such a case, any non-NaN result was regarded as correct.