[src/vasprintf.c] Bug fix for the l length modifier with FP type.
The l length modifier is now supported with the a, A, e, E, f, F, g, G
conversion specifiers, as required by the ISO C standard on printf and
other formatted output functions. As specified, it is just ignored.

For that, to improve code maintainability, the CASE_LONG_ARG() macro
was simplified:
  * The c and s conversion specifiers are always supported by this
    macro (this is a potential fix for implementations with incomplete
    wide-character handling, without any guarantee to work).
  * A MPFR_RET_NEVER_GO_HERE() for unsupported conversion specifiers
    has been added (if specinfo_is_valid() is consistent with this
    macro, one should effectively never reach this case).