[tests/tsprintf.c] Correction for %la, %lA, %le, %lE, %lf, %lF, %lg, %lG
Reverted commit 2b60707d22717e0f582be5de680d7fa7004e2512 and added
tests for the missing %lE, %lF, %lG in native_types, protected by
a #if (see commit ecd3279f3dfd281fc8e51aac7509b76b27497a6c) as they
may need an ISO C99 compiler.

Note: The test of %la and %lA was incorrect because the first digit
is unspecified (and really depends on the C library). A test could
be added in the future, perhaps also with a configure test, because
the compiler may support ISO C99 but not the C library (there was
an issue like that at least with old versions of Solaris).