[src/vasprintf.c] Consistency with the non-support of %F by GMP.
The F conversion specifier as defined by the C standard is not
supported by the GMP formatted output functions because it is
already used as the type specifier for mpf_t (which comes before
the conversion specifier). Let's take this fact into account.

Note: this conversion specifier is still allowed with MPFR numbers
(e.g., "%RF") to avoid breaking the compatibility (it has never
been an issue).