Support ChangeLog regeneration from a Git working tree.
* Added tools/gitlog-to-changelog file from Gnulib.
* Makefile.am: added a rule for the Changelog file.

Needed as the GNU MPFR repository has been converted to Git.

The '%B%n' format string is used rather than the default '%s%n%b%n'
because Git's commit message convention is not always followed
(in particular before the conversion of the repository to Git).

Note: In the future, if there is any change to do to a Git commit
message (typo, etc.), the --amend option could be used.