[mbench/timp.h] Bug fix for "make rt" on 32-bit machines. Reindent.
For the code used when USE_CLOCK_MONOTONIC is defined ("make rt"):
If both time_t (type of tv_sec) and long are 32-bit types, there was
an overflow in the timp_rdtsc() macro, due to the multiplication of
the tv_sec (coming from CLOCK_MONOTONIC, thus typically larger than 5)
by the value 1000000000 (≈ 2^32 / 4.3) of type unsigned long, because
this multiplication would be done on 32 bits. Changed the type of the
constant to unsigned long long so that the operations are done with
the target type unsigned long long (at least).