Fixed tsprintf %a tests when GMP and libc produce different outputs.
With some 32-bit mingw-w64 implementations, the GMP and libc
formatted output functions produce different outputs for %a (this
is not completely specified by the C standard). In the MPFR tests
introduced by commit 22db634294c27eabbc0c55b04267d3ab5bc9a648 on
2022-05-19, the comparison is done between the MPFR result, which
is based on GMP, and the libc result; hence a failure.

So, changed the 4 concerned tests of tests/tsprintf.c to compare
the MPFR result with the GMP result (for the other tests on the
C native types, it is still better to compare MPFR with libc).

Bug report:
  https://sympa.inria.fr/sympa/arc/mpfr/2022-06/msg00004.html