[doc/mpfr.texi] Updated @m{T,N} macro, with comments.
The @m{T,N} macro is used for math formatting and textual equivalent
(human-readable text). However, it was generating emphasis in HTML,
which was unexpected. This commit removes the use of @math for the
Info and HTML output (for Info, this does not change anything, and
for HTML, this removes the incorrect emphasis). It also updates the
documentation and adds a comment about the history of this macro and
various issues related to it.