[tools/update-version] Check for failing replacements.
This script currently fails due to a change done on 2018-09-07 in
doc/mpfr.texi (commit 84796030c7c732e8e66e5e650ec929c541dd207f /
SVN r13153), and this issue remained unnoticed until now,
in particular due to the absence of checking for failing
replacements.

The penultimate non-patchlevel release was done on 2017-12-25, i.e.
before this change, and no patchlevel releases have been done yet
for the latest release, so that no current releases are affected by
this issue, but the 4.1.1 release candidate is affected.