[tools/update-version] Fixed replacements in "doc/mpfr.texi".
The change done in commits 4885b57add174f390eff29abe14d12fb5ebd6a61
and 7eda345244f503359d14cd5bae1ed42865d39ee8 was incorrect as there
are 2 identical replacements to do in "doc/mpfr.texi", not just one.
The previous code changed lines separately. But for the check of
failing replacements, a global change on the full file is needed;
thus the 2 identical replacements were expected to be done at the
same time, so that the "g" modifier is needed on the regexp.