Get rid of the _mulx_u64 intrinsic.
It was found that this intrinsic has issues with ICC, Cygwin, and more
generally, the Microsoft toolchain; even MPFR's strong detection at
configure time (an AC_RUN_IFELSE test) was not sufficient. Moreover,
_mulx_u64 is useless as umul_ppmm from mpfr-longlong.h does the same
thing from the spec point of view (and thus was used as the fallback).
So, if there is a reliable way to use the mulx instruction whenever
possible (including via compiler optimization and the use of __int128
when available, as done by GCC's _mulx_u64 implementation), this
should be done in the umul_ppmm definition. Alternatively, we could
redefine umul_ppmm for the use of __int128 when possible.

Context and references:
  https://sympa.inria.fr/sympa/arc/mpfr/2022-11/msg00008.html
  https://sympa.inria.fr/sympa/arc/mpfr/2022-11/msg00027.html
  https://sympa.inria.fr/sympa/arc/mpfr/2022-11/msg00030.html
  https://sympa.inria.fr/sympa/arc/mpfr/2022-11/msg00031.html
  https://stackoverflow.com/a/54815033/3782797

* src: corrected the umul_hi(h, x, y) macro (the only code that was
  using _mulx_u64) and moved it from invert_limb.h and invsqrt_limb.h
  to mpfr-impl.h.
* tests/tversion.c: removed the output info about _mulx_u64.
* acinclude.m4: removed the check for _mulx_u64.