[configure.ac] Added a comment about a possible AM_PROG_AR failure.
The reason is that on some platforms, GMP chooses a non-default mode
(a.k.a. ABI). MPFR has some workaround thanks to GMP_CC and GMP_CFLAGS
provided by GMP, but this may not be sufficient, e.g. on power-aix. In
such a case, the user needs to either force GMP to use the default mode
or provide an AR variable for MPFR's configure; see INSTALL file.