[tests/tsprintf.c] Updated comment about the possible test failure.
This issue started to appear with OpenBSD 7+ too, which no longer
supports the standard %n printf specifier. This is detected by
GMP's configure, so that HAVE_VSNPRINTF is not defined. If GMP is
built with HAVE_VSNPRINTF manually defined, this would would make
the failure with %a/%A disappear, but would trigger failures due to
the %n test in several MPFR tests.