Revert "[tests/tsprintf.c] Updated comment about the possible test failure."
This reverts commit b5f229d1c4019800fdbd83a7ecc9eda12663d38e.

The test of the %n printf specifier with the mpfr_*printf functions
is actually not an issue since we handle %n in vasprintf.c (so the
OpenBSD C library will never see it). That's why with a GMP compiled
under OpenBSD 6.9 (thus with HAVE_VSNPRINTF defined), the MPFR tests
still succeed after an upgrade of the machine to OpenBSD 7.2.