[tests/tget_ld_2exp.c] No longer check the smallest positive subnormal.
The tget_ld_2exp failure was apparently due to a bug in QEMU, which
is not aware of the m68k extended-precision format. The MPFR code is
correct (the HAVE_LDOUBLE_IEEE_EXT_BIG macro is not used yet, and if
it is used in the future, we could decide that it corresponds to the
m68k extended-precision format, as the x86 extended-precision format
seen in practice is little endian, i.e. HAVE_LDOUBLE_IEEE_EXT_LITTLE
is set). There's now check_qemu_m68k_bug() about this QEMU bug.