[tests/tget_ld_2exp.c] Removed check_smallest_subnormal completely.
This code was commented out in the previous commit. Let's remove it
completely because it was buggy for HAVE_LDOUBLE_IEEE_QUAD_*, as this
has been noticed by Joseph Myers:
  https://sympa.inria.fr/sympa/arc/mpfr/2022-12/msg00052.html

Moreover, the HAVE_LDOUBLE_IEEE_EXT_BIG interpretation is not decided
yet and could be used for m68k, with "e2 = -16446" instead of -16445.