[tests/tget_ld_2exp.c] Do not always run check_qemu_m68k_bug().
First, the check_qemu_m68k_bug() code was valid only with 15-bit
long double exponents or more, thus not on every architecture.
Moreover, the MPFR tests are not there to detect bugs in the native
floating-point arithmetic. So this check is now done only when
bug20180904() fails (as this QEMU bug is now the probable cause of
this failure) in order to instruct the user that this is probably
not a MPFR bug.

TODO: Make the code C90 compatible.