[src/zeta.c] Attempt to correct/clarify my comment about mpfr_sinpi.
Note: My previous explanation was implicitly based on the fact that
tiny arguments s were already taken into account. But if the precision
is really huge (see the condition), this is not the case, so that the
MPFR_EMAX_MAX lower bound on the precision may have been incorrect.
Anyway, as noticed by Paul, one can use the fact that one would have
|s| less than the minimum representable positive number (i.e. without
needing the filtering of tiny arguments).