[tests.c] In bad_cases(), handle the inex_inv case better.
The "f exact while f^(-1) inexact" case is actually almost the same
as the "inverse doesn't match" case. So, handle it in the same way,
avoiding a fatal error (which would be bad for the non-developers):
a few occurrences of this case should be regarded as normal, and if
this occurs too often, this will be detected like the other case
when the MPFR_CHECK_BADCASES environment variable is set.