[tests/tui_pow.c] Added mpfr_ui_pow_ui test whose result is +Inf.
It triggers the bug fixed in 99e17f3d31211af2f9913059bc529d731fed1513
(handle the +Inf case in the Ziv iteration of mpfr_ui_pow_ui).

The tlog10.c bug20230213() test was triggering it too, but it depends
on the mpfr_log10 implementation, which could change in the future.