[src/vasprintf.c] Added MPFR_ASSERTD and code cleanup in regular_fg().
Commit 70441828717299d43afd5844410ac9b523cad00e added checks for
potential integer overflows (in practice, for platforms where
mpfr_intmax_t has 32 bits), but in case of avoided integer overflow
thanks to some "if" test, the code would go on with incorrect values
instead of returning an error. However, with a careful analysis of
the code, one can prove that these integer overflows were actually
not possible thanks to the change of some types by the mentioned
commit (see added comments).

Also removed the "spec.prec > 0" from the "if" test since this is
always true in this branch of the code.