[src/vasprintf.c] Added a comment for regular_fg().
This explains why the code is correct. As a summary, in all branches,
np->fp_size may be incorrect if there is no output, due to the use of
mpfr_get_str_wrapper, but in such a case, what matters is that the
sum np->fp_size + np->fp_trailing_zeros is correct.