Possible stack overflow in mpfr_rec_sqrt: added testcase + FIXME.
The issue is due to many allocations in the stack on cases that are
very hard to round (testcase given by Fredrik Johansson) with a not
too large destination precision: it is not possible to free the data
allocated in the previous iterations.

Possible solution: Use a standard Ziv loop (why is this not done?),
as it increases the working precision multiplicatively instead of
additively, so that there would be much fewer iterations. This would
also make the code much faster in such cases.

Note: This issue was not detected by bad_cases() in tests/trec_sqrt.c
because one needs a much larger precision than the ones tested in the
chosen parameters. In the testcase provided by Fredrik Johansson, the
destination precision is 79337, and a working precision up to 123776
is needed.

* src/rec_sqrt.c: added a FIXME about a possible stack overflow.
* tests/trec_sqrt.c: added a bad_cases() call with larger precisions
  in order to trigger the stack overflow on current Linux machines,
  at least with 64-bit limbs.