Added MPFR_LOG_ALLOCA_F log type to log the alloca() amounts.
Added associated environment variable MPFR_LOG_ALLOCA.

Note about the implementation: In each function where a TMP_ALLOC()
is done, a variable __gmpfr_log_alloca_size needs to be declared.
In order to ensure that only one variable is declared (even when
several markers are used), we do this in MPFR_LOG_FUNC. This means
that the function needs to be logged with MPFR_LOG_FUNC (anyway,
this is a good thing as this allows one to know when the function
is called and when it returns, which is important information for
alloca() logging, since the corresponding memory is freed when the
function returns).

* doc/README.dev: document the MPFR_LOG_ALLOCA environment variable.
* src/mpfr-impl.h: define macro MPFR_LOG_ALLOCA_F and in MPFR_LOG_FUNC,
  declare and init the __gmpfr_log_alloca_size variable (also use it
  to avoid compiler warnings in functions that do not use TMP_ALLOC).
* src/mpfr-gmp.h: if logging is enabled, update __gmpfr_log_alloca_size
  when an alloca() is done.
* src/logging.c: take MPFR_LOG_ALLOCA into account for mpfr_log_type.
* src/*.c: add MPFR_LOG_FUNC logging in functions that use TMP_ALLOC
  (when this is not already done).