mpfrlint: detect a Ziv loop that does not use the MPFR_ZIV_* macros.
Such a code may give a bug like the one in rec_sqrt.c, found and fixed
in 2023-04.

This is a heuristic; it assumes GNU style with correct indentation.
This can be silenced by using MPFR_ZIV_NEXT in a comment in the loop.
A false positive may actually be a bad indentation, which should be
fixed anyway.