[TODO] Added a TODO.
Check potential issues between aliases and the reuse of arguments
(typical case: mpfr_fms): if the detection of reuse is done by a
comparison on the pointers to the MPFR numbers (as sometimes done),
this detection will fail since these pointers are different though
these numbers share their significand.

Note: It seems that the reuse in mpfr_fms may propagate to mpfr_add1
and mpfr_sub1. Fortunately, add1.c and sub1.c do the reuse detection
by comparing the pointers to the significands, so that the code is
working. But this needs a closer check and some testing. Comments
will also be added to warn against these issues.