Added treldiff.c test, assuming that mpfr_reldiff behaves as documented.
Basically, mpfr_reldiff is tested with all special numbers and some
regular ones, with various precisions and all the rounding modes
(including MPFR_RNDF since in practice, the implementation cannot
exploit it to give failures). The result is just compared with the
basic algorithm (as documented), which does not use any optimization
(i.e. it does not treat special numbers in a particular way): the
formula is just computed using the precision of the destination and
the provided rounding mode for all operations (thus, in general,
this mpfr_reldiff function is not correctly rounded).

This test currently fails because of several bugs on special cases
(the behavior follows neither the documented algorithm nor the usual
rules for the special cases).