[tests/texp10.c] Updated comments about bug20230427 (bug_exp10_2).
A testcase for the 2nd mpfr_exp10 bug (actually in mpfr_pow_general)
has been given with the merge of the bug_exp10_2 branch: bug20230427.
So, updated the TODO about hardcoded tests. A hardcoded test for the
first bug (fixed in b62966df913f73f08b3c5252e1d0c702bc20442f) still
needs to be given, in particular for a port to the 4.2 branch, which
will not have the (currently buggy) ofuf_thresholds tests.

Also detailed the comment for bug20230427.