Added a '\0' test for mpfr_inp_str.
After ensuring that isspace(0) is false, make sure that '\0' is
handled as a normal non-whitespace character.

Also note that inp_str.dat does not end with a newline character
on purpose, in order to check that a number at the end of the file
(with no whitespace after it) is handled correctly.