[inp_str.c] mpfr_inp_str: fixed the handling of '\0' in the word.
When '\0' is not a whitespace character (which is almost always the
case in practice), it was not handled correctly. For instance, if the
word is the sequence { '1', '\0', '2' }, the string "1" was passed to
mpfr_set_str because '\0' is regarded as a terminating null character,
and one was getting a valid number while '\0' in a word is necessarily
invalid. This is fixed by putting '*' at the beginning of the string,
which is thus necessarily invalid.

This fixes the failure after c9523bc6fb11c8eefc75524e947e5eb81e682675.