Fixed mpfr_strtofr bug in parse_string when isspace(0) is true.
When isspace(0) is true (allowed by ISO C for non-"C" locales), '\0'
would be skipped if part of the leading whitespace. But it must be
regarded as the terminating null character, from the definition of
a string (before the notion of whitespace is involved). So let's
stop explicitly on '\0'.

Note: in such locales (Mutt's lib.h suggests that some systems are
affected, but this was in 1998), this is a security vulnerability,
because characters after the terminating null character are read to
determine the result.