[manual] Added warnings about whitespace for mpfr_strtofr & mpfr_inp_str.
In short, it is probable that '+', '-' and the decimal-point character
will never be regarded as whitespace, because standard functions like
strtod would obviously not work as wanted. But there is no guarantee
for '@' and '_' (as having them as whitespace would not affect the
standard functions), which may appear in strings representing MPFR
values but could be read as being whitespace.