[configure.ac] More checks for the support of old dtags.
There is an attempt to use the old dtags when LD_LIBRARY_PATH is
defined in order to avoid issues with the testsuite when this path
contains some compatible version of the MPFR library. However, the
old dtags do not work everywhere at run time, even when they appear
to work at link time (via -Wl,--disable-new-dtags).

So, let's restrict to *-*-linux-gnu hosts. Unfortunately, this is
not sufficient for Android/Termux, where config.guess currently
identifies the host as GNU/Linux (matching *-*-linux-gnu) instead of
Android, but the old dtags do not work at run time. So, let's also
exclude Termux explicitly (a test to exclude all Android hosts would
be better, though).

Note: in most cases, LD_LIBRARY_PATH is not set, so that there should
be no issues in practice.