About the .POSIX special target, mention the AIX case (this is OK).
From tests on cfarm119.cfarm.net (AIX 7.3):

Due to the Automake issue with .POSIX, under AIX the "make" command
outputs the following warning:
  warning: .POSIX directive is not first non-comment line.

Its man page says that /usr/ccs/lib/posix.mk provides the default
rules, but the behavior is not clear when ".POSIX:" is not the
first non-comment line.

I could see that using ".POSIX:" changes $(CC) from "cc" to "c89"
via "CC=c89" in this file. I've checked that this does not change
$(CC) to c89 when ".POSIX:" appears after the makefile sets CC.