[vasprintf.c] Attempt to fix various issues (but GMP is buggy for %c on 0).
The use of strncpy and strncat is unnecessary, and even incorrect
when the buffer contains a non-terminating null character, which
is possible when %c is used with the value 0. Use memcpy instead.

2 MPFR_ASSERTD occurrences involving strlen are incorrect for the
same reason (a null character due to %c with the value 0 would be
regarded as terminating). So let's remove them.

The use of mpfr_free_str in the FLUSH macro is also incorrect for
the same reason. Instead, let's call mpfr_free_func directly, so
that we can replace strlen(s) (where the first null character is
regarded as terminating) by the return value of gmp_vasprintf.
However, this return value is incorrect, which still makes the
test fails (even if the allocation function ignores the size
parameter, e.g. by setting "tests_memory_disabled = 1;" at the
very beginning of the main() function of the tsprintf.c test).

So the GMP bug needs to be fixed:
  https://gmplib.org/list-archives/gmp-bugs/2023-December/005420.html