mpfr_compound_si: updated comments and fixed the code again.
Before commit d390bf2b2ff645ff129da8b977331b23d91a0f5a, one had
  e = (precu - prec >= e) ? 1 : e + 1 - (precu - prec);
for the test
  if (mpfr_min_prec (t) <= py + 1 && prec - py >= e + 1)
which uses "the approximation t is at most 2^e ulp(t)".

Due to the rename of e (old code) to e3 in the new "if (e < precu)"
branch, the approximation t is now at most 2^e3 ulp(t) there. Thus
e needs to be changed to e3 in the above test, i.e.:
  if (mpfr_min_prec (t) <= py + 1 && prec - py >= e3 + 1)

Note: commit 8c8aa82aefc69d9960b2d59df695e235c88f3b21 was not the
right fix. The issue was the e >= precu case, which got fixed in
commit f0248d447059253e878b6cc26364a4a4a0366044. Still, the meaning
of "loop again" in this case is unclear as this is not what the code
does.