MPFR manual: updated mpfr_compound description about special cases.
These are specifically the mpfr_pow rules, which extend the usual
rules. The other special cases were the same as the usual rules
for mpfr_compound_si, but with the new mpfr_compound function,
the second argument of compound(x,n) can now be ±Inf or NaN. Thus
compound(0,y) gives 1 even when y is ±Inf or NaN, following the
pow rules (contrary to the usual rules).