[tests/tcompound.c] Additional tests for compound(-1,y).
In particular, test special values of y and check the flags. This
shows a bug for compound(-1,-Inf), which sets the divide-by-zero
flag, even though y is an infinity. Note that the IEEE 754 standard
specifies: "pow (±0, −∞) is +∞ with no exception".

This also shows a bug in flags_out(), used for the tests.