version-ext.sh: handle a particular case due to merged Git branches.
It is possible that $gitb contains multiple branches. This may be
caused by a branch B that was merged into a branch A while it was
in sync with branch A, i.e. before additional commits to branch A.
In such a case, a merge commit isn't even generated, and the status
is unfortunately ambiguous.

Let us handle only the practical case where branch A is master by
assuming that the expected branch is master.

Note: if such an issue occurs in other cases, the user may want to
avoid it be deleting local branches that are no longer needed.

This was occurring at commit 4f2d2d5ee1453f8a376025ff96a8165fb9ca8911
(with the merge of the compound branch into master).