[sin_cos.c] Removed buggy dead code in sin_bs_aux.
Vincent Lucarelli found a bug in sin_bs_aux for the case p = 0:
S0 is set to 1 instead of 0. But as he noticed, before the call
to this function, there is a test that eliminates this case.
See https://sympa.inria.fr/sympa/arc/mpfr/2024-04/msg00004.html

Rather than fixing this latent bug, let's replace this dead code
by a MPFR_ASSERTD that checks that p > 0. In the spec of this
internal function, replaced "0 <= X=p/2^r" by "0 < X=p/2^r" to
reflect this change.

Note: the code was introduced in 2009
(commit 094393538587397f19d5dbd439f5c66ebc7e229b),
and on 2012-07-05, a coverage test gave 290 / 299 lines
for this file.