mpfrlint: check "zeroes"; associated fixes in various files.
Contrary to "zeroed" and "zeroing", which are sometimes used, "zeroes"
as a verb (third-person singular) is rare in practice. Most often, this
word appears as a noun (plural of "zero") instead of "zeros", which is
the most common spelling and what we should use. So let's warn about it
(in case of false positive, we should add an exception with "grep -v").

Note: that's why in SCOWL (Spell Checker Oriented Word Lists)[*], "zero"
is in english-words.10 and "zeros" in english-words.20, but "zeroes" is
only in *variant_2-words.20 (as a variant of the plural of "zero") and
special-hacker.50 (as a verb).

→ 10 fixes in various files (all the occurrences were the noun).

[*] http://wordlist.aspell.net/