[troot.c] For the output of timings, use a fixed-point format (%f).
For the output of timings (2 arguments: precision and k), using
the %e format makes comparison more difficult when the exponents
for mpfr_rootn_ui and for exp(1/k*log(x)) are different. It is
better to use a fixed-point format (even though some trailing
digits are not significant). Also ensure that both timings are
aligned (assuming they are < 1000s).