MPFR manual: removed partly incorrect text about mpfr_rootn_ui.
Removed "the memory usage of mpfr_rootn_ui is also linear in the
argument k", which is true only for k up to 100 (hardcoded constant).
Moreover, this was in a section on transcendental functions while
mpfr_rootn_ui is an algebraic function. And the current choice may
be up to discussion and might change in the future (if memory usage
is an issue for k = 100, this may mean that we made a wrong choice).