[trigamma.c] Added a FIXME and a MPFR_LOG_MSG about an overflow.
The result is computed with MPFR_RNDN, which gives an infinity
instead of a finite value in MPFR_RNDZ or equivalent.

Note: mpfr_check_range currently assumes that the underflow/overflow
detection in the Ziv loop is correctly handled, e.g. by calling
mpfr_underflow or mpfr_overflow (mpfr_check_range just handles
overflows triggered when a finite approximation is rounded).

The case is currently tested by the generic tests (x = 0+, MPFR_RNDD),
and an inexact +inf is returned, which should never be possible with
this rounding mode. Such a check should be added to the generic tests.